Flatten PermissionsMatrix into single file
