--- src/components/form/PermissionsMatrix/index.js
@@ -1,2 +1,0 @@
-export { PermissionRow } from './PermissionRow'
-export { PermissionsMatrix } from './PermissionsMatrix'
