+++ src/components/form/index.js
@@ -2,7 +2,7 @@ export { FormFields, FormGroup, FormRoot, FormRow } from './Form'
 export { FormBoolController } from './FormBoolController'
 export { FormController } from './FormController'
 export { FormField } from './FormField'
-export { PermissionsMatrix } from './PermissionsMatrix'
+export { PermissionsMatrix } from './PermissionsMatrix.jsx'
 export { SubmitButton } from './SubmitButton'
 export { TeamAddForm } from './TeamAddForm'
 export { UserInvitationForm } from './UserInvitationForm'
